only authorized users are able to delete threads

+++ app/Http/Controllers/ThreadsController.php
@@ -118,6 +118,10 @@ class ThreadsController extends Controller
      */
     public function destroy (Channel $channel, Thread $thread)
     {
+        if ($thread->user_id != auth()->id()) {
+            abort(403, 'You are not permitted');
+        }
+
         $thread->delete();
 
         if (request()->wantsJson()) {
